chore: disable line length for headings
When I began adding standarization to some of our projects with changelogs,
I noticed our line length rules were complaining about changelog headers,
which contain alot of content (IE: long compare URLs). I talked to @yokuze
and we came to the following conclusion:

We should disable the line length check for headings because:

Any increased line length for headings would be an arbitrary
value that we would probably violate again in the future anyway
We can’t wrap headings to multiple lines like we could with
other pieces of content, so they will always be long if they
include a link.