chore: reconfigure markdownlint to use CLI and NPM scripts
This commit installs markdownlint-cli, removes other markdownlint libraries, and switch
to an NPM script to kick of markdown linting. We chose to move to a NPM script due to
the fact that we are looking to move away from grunt. Also grunt-markdownlint declares
minor resolved versions of markdownlint proper as a dependency, and therefore we cannot
rely on the correct version of markdownlint being resolved against the correct version
of our markdowlint config. Grunt-markdownlint also does not export extends for
markdownlint configurations.