feat: Add acknowledgment for a release that has no changelog entries
In the event that no changelog entries are found for a release, we add a
line acknowledging this, so that there is content associated with the
changelog header for this version.

This phenomenon can happen since a release may be made that consists of
commits that are not included as changelog entries, like `build`,
`chore`, or `docs`.

Previously, the release process would abort if there were no changelog
entries. This change allows us to still update the changelog with the
version release when only internal/documentation commits make up the
release.